requirements.txt
PyGithub==1.26.0
argparse==1.4.0
decorator==4.0.9
ipython==4.1.2
ipython-genutils==0.1.0
pathlib2==2.1.0
pickleshare==0.7.2
prboard==1.0a0
pyreadline==2.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
